added vim-utl for text-based URL traversal support
This sets the default web browser to Safari.  That really needs
to be guarded by an OS X check.  Also, I hit a problem where
if I used the fragment in the link construction, it would do
the wrong thing and fill in some crud if no #fragment was
specified.  I didn't drill into that and just skipped the
fragment for now.  This should be fixed if I want full
link traversal semantics to work on web pages.